working on transparency

+++ RvtVa3c/Va3cScene.cs
@@ -28,7 +28,7 @@ namespace RvtVa3c
       [DataMember] public int emissive { get; set; } // 1
       [DataMember] public int specular { get; set; } //1118481
       [DataMember] public int shininess { get; set; } // 30
-      [DataMember] public int opacity { get; set; } // 1
+      [DataMember] public double opacity { get; set; } // 1
       [DataMember] public bool transparent { get; set; } // false
       [DataMember] public bool wireframe { get; set; } // false
